% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCPRData.R
\name{queryCPRData}
\alias{queryCPRData}
\title{Queries CPR data}
\usage{
queryCPRData(forecast.type, gen.plan = NULL, username = NULL,
  password = NULL, driver.name = "SQL Server")
}
\arguments{
\item{forecast.type}{character indicating the desired forecast. Users can generate a table of available
options using \code{queryCPRForecasts()}.}

\item{gen.plan}{character indicating the desired forecast vintage. Users can generate a table of available
options for a given forecast using \code{queryCPRGenPlans()}, or leave as NULL to select most recent vintage.}

\item{username}{character for connecting to the database. Leave username and/or password as NULL to use
Windows Authentication via a trusted connection}

\item{password}{character for connecting to the database}

\item{driver.name}{character for the database driver to be used}
}
\value{
A data frame, with a variable number of columns corresponding to the layout of the underlying CPR XML files.
}
\description{
Queries TVA's Characteristics of Power Resources (CPR) database for planning assumptions, 
including fuel prices and generation unit characteristics.
}
\details{
Resource Planning and Strategy maintains
}
\examples{
# query list of forecasts
forecasts = queryCPRForecasts()

# query data for the most recent forecast vintage
gas.price = queryCPRData(forecast.type = 'Henry Hub Natural Gas')

# obtain gen plans for a given forecast type
gen.plans = queryCPRGenPlans(forecast.type = 'Henry Hub Natural Gas')

# query data for a given forecast and gen plan
gas.price.spsp = queryCPRData(forecast.type = 'Henry Hub Natural Gas', 
                              gen.plan = 'FY18 Strategic PSP')
}
