% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCPRForecasts.R
\name{queryCPRForecasts}
\alias{queryCPRForecasts}
\title{Queries available forecasts}
\usage{
queryCPRForecasts(username = NULL, password = NULL,
  driver.name = "SQL Server")
}
\arguments{
\item{username}{character for connecting to the database. Leave username and/or password as NULL to use
Windows Authentication via a trusted connection}

\item{password}{character for connecting to the database}

\item{driver.name}{character for the database driver to be used}
}
\value{
A three column data frame, including ForecastID, TemplateID, and Name for each forecast
}
\description{
Queries TVA's Characteristics of Power Resources (CPR) database for available forecast types, 
including fuel prices and asset characteristics.
}
\details{
This function creates a connection to the CPR database, which is maintained by Resource Planning and
Strategy, and contains the forecast data used to feed long-term generation planning models.
}
