% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCPRGenPlans.R
\name{queryCPRGenPlans}
\alias{queryCPRGenPlans}
\title{Queries available gen plan vintages}
\usage{
queryCPRGenPlans(forecast.type, username = NULL, password = NULL,
  driver.name = "SQL Server")
}
\arguments{
\item{forecast.type}{character indicating the desired forecast. Users can generate a table of available
options using \code{queryCPRForecasts()}.}

\item{username}{character for connecting to the database. Leave username and/or password as NULL to use
Windows Authentication via a trusted connection}

\item{password}{character for connecting to the database}

\item{driver.name}{character for the database driver to be used}
}
\value{
A two column data frame, including ForecastTitle and UploadDate
}
\description{
Queries TVA's Characteristics of Power Resources (CPR) database for available gen plans, 
including monthly gen plans, and strategic and budget power supply plans, for a given forecast type.
}
\examples{
# query list of forecasts
forecasts = queryCPRForecasts()

# obtain gen plans for a given forecast type, listed in \\code{forecasts}
gen.plans = queryCPRGenPlans(forecast.type = 'Henry Hub Natural Gas')
}
